Fix Diff Issue
The diff function has a problem with complex custom fields because
they are unmarshalled as interface{} values, either map[string]interface{}{}
or []interface{}{}. If you compare an unmarshalled object to one modified
in code, it incorrectly returns true.  Added a string comparsion to
short circuit the Diff when they are truly equal.

-r mhupman