Add Custom Field Deserializer
Adds deserializer for custom fields that provides concrete types for custom
fields.  This should improve the diffing logic as types will be compared to
other concrete types instead of interfaces.  Additionally, this modifies the
LocationService to return Hydrated locations. Also modifies the differ to break
out of the shitty logic if a location is hydrated.

r=mhupman