Change CustomFieldCache to CustomFieldManager
CustomFieldCache is now CustomFieldManager and contains Getters and Setters for
custom fields of all types.  CustomFields now implement the CustomFieldValue
interface so that the Getters and Setters are typed.  This will ensure that
diffing against hydrated locations will produce acurate results. Additionally,
adds getters and setters for Single and Multi Option custom fields as well as a
method to determine if a specific value is set on the field without having to
know the OptionIds.

r=mhupman